backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic[email]>=2.6.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
httpx>=0.26.0
aiofiles>=23.2.0
selenium>=4.18.0
webdriver-manager>=4.0.1
markdownify>=0.11.6
beautifulsoup4>=4.12.3
requests>=2.31.0
Pillow>=10.2.0
imageio>=2.37.0
imageio-ffmpeg>=0.6.0
psutil>=5.9.8  # System resource monitoring (CPU/memory/disk/network)

# Data Analysis & Excel Processing (for Excel Analysis Workflow)
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # Excel file reading/writing
matplotlib>=3.7.0  # Chart generation

# Database
sqlalchemy[asyncio]>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0  # bcrypt 4.0+ 与 passlib 1.7.4 不兼容

# PDF Processing & AI
pypdf>=5.0.0
google-genai>=1.55.0  # 统一使用新版 SDK，支持聊天和图片生成
dashscope>=1.17.0  # 阿里云 DashScope SDK（用于通义千问）
openai>=1.12.0  # OpenAI SDK（用于 OpenAI 和 Ollama 服务）
tiktoken>=0.5.0  # Token counting library（用于 Ollama 服务）
cachetools>=5.3.0  # Caching utilities（用于 Ollama 服务）
google-cloud-aiplatform[agent_engines,adk]>=1.38.0  # Vertex AI SDK（用于 Virtual Try-On、高级图像功能和 Agent Engine）
google-adk>=0.1.0  # Agent Development Kit（用于 Agent Engine 高级功能）

# Document Embedding / RAG
scikit-learn>=1.3.0
# Note: numpy is now listed under Data Analysis section above

# .env file support
python-dotenv>=1.0.0

# PostgreSQL driver (optional, uncomment to use PostgreSQL)
psycopg2-binary

# Task Queue (Redis-based Worker Pool)
redis>=5.0.0

# MCP (Model Context Protocol)
mcp>=0.9.0  # MCP SDK for tool integration

# Cloud Storage Providers
oss2>=2.18.0  # 阿里云 OSS SDK
cos-python-sdk-v5>=1.9.0  # Tencent Cloud COS
google-api-python-client>=2.0.0  # Google Drive API
google-auth-httplib2>=0.1.0  # Google Auth HTTP
google-auth-oauthlib>=1.0.0  # Google OAuth
boto3>=1.26.0  # AWS S3 and S3-compatible storage
